Started with card table

+++ src/runtime/memory/generational/utils.h
@@ -9,7 +9,7 @@
 #define INCREMENT_GENERATION(object) \
     do {                                 \
         uint8_t new_generation = GET_GENERATION(object) + 1; \
-        uint8_t * forwading_ptr = GET_FORWARDING_PTR(object);                         \
+        uint64_t * forwading_ptr = GET_FORWARDING_PTR(object);                         \
         object->gc_info = (void *) (((uint64_t) forwading_ptr) | ((uint64_t) (new_generation << 61)));  \
     } while(false);
 #define GET_GENERATION(object_ptr) ((uint8_t) object_ptr->gc_info >> 61)
